Fix auth tests
